ngubah main.js buat ngarah ke lib db
