+++ lib/DB_Adapters/cloudDBAdapter.js
@@ -1,6 +1,6 @@
-import got from 'got';
+import got from 'got'
 
-const stringify = obj => JSON.stringify(obj, null, 2);
+const stringify = obj => JSON.stringify(obj, null, 2)
 const parse = str => JSON.parse(str, (_, v) => {
     if (
         v !== null &&
@@ -9,21 +9,20 @@ const parse = str => JSON.parse(str, (_, v) => {
         v.type === 'Buffer' &&
         'data' in v &&
         Array.isArray(v.data)) {
-        return Buffer.from(v.data);
+        return Buffer.from(v.data)
     }
-    return v;
-});
-
-export const cloudDBAdapter = class CloudDBAdapter {
+    return v
+})
+class CloudDBAdapter {
     constructor(url, {
         serialize = stringify,
         deserialize = parse,
         fetchOptions = {}
     } = {}) {
-        this.url = url;
-        this.serialize = serialize;
-        this.deserialize = deserialize;
-        this.fetchOptions = fetchOptions;
+        this.url = url
+        this.serialize = serialize
+        this.deserialize = deserialize
+        this.fetchOptions = fetchOptions
     }
 
     async read() {
@@ -34,11 +33,11 @@ export const cloudDBAdapter = class CloudDBAdapter {
                     'Accept': 'application/json;q=0.9,text/plain'
                 },
                 ...this.fetchOptions
-            });
-            if (res.statusCode !== 200) throw res.statusMessage;
-            return this.deserialize(res.body);
+            })
+            if (res.statusCode !== 200) throw res.statusMessage
+            return this.deserialize(res.body)
         } catch (e) {
-            return null;
+            return null
         }
     }
 
@@ -50,8 +49,10 @@ export const cloudDBAdapter = class CloudDBAdapter {
             },
             ...this.fetchOptions,
             body: this.serialize(obj)
-        });
-        if (res.statusCode !== 200) throw res.statusMessage;
-        return res.body;
+        })
+        if (res.statusCode !== 200) throw res.statusMessage
+        return res.body
     }
 }
+
+export default CloudDBAdapter
